// src/models/post.d.ts
export interface Welcome {
    data: WelcomeData;
    meta: Meta;
}

export interface WelcomeData {
    id:         number;
    attributes: PurpleAttributes;
}

export interface PurpleAttributes {
    title:       string;
    description: string;
    content:     string;
    createdAt:   Date;
    updatedAt:   Date;
    publishedAt: Date;
    slug:        string;
    cover:       Cover;
}
// ...
export interface Meta {
}